Refactor the recurring callback support.
