+++ substratevm/src/com.oracle.svm.core/src/com/oracle/svm/core/heap/ReferenceHandlerThread.java
@@ -36,7 +36,7 @@ import com.oracle.svm.core.SubstrateOptions;
 import com.oracle.svm.core.Uninterruptible;
 import com.oracle.svm.core.feature.AutomaticallyRegisteredFeature;
 import com.oracle.svm.core.feature.InternalFeature;
-import com.oracle.svm.core.thread.ThreadingSupportImpl;
+import com.oracle.svm.core.thread.RecurringCallbackSupport;
 import com.oracle.svm.core.thread.VMThreads;
 import com.oracle.svm.core.util.VMError;
 
@@ -76,7 +76,7 @@ public final class ReferenceHandlerThread implements Runnable {
 
     @Override
     public void run() {
-        ThreadingSupportImpl.pauseRecurringCallback("An exception in a recurring callback must not interrupt pending reference processing because it could result in a memory leak.");
+        RecurringCallbackSupport.suspendCallbackExecution("An exception in a recurring callback must not interrupt pending reference processing because it could result in a memory leak.");
 
         this.isolateThread = CurrentIsolate.getCurrentThread();
         try {
